Fixed bug with arbitrary resolution grid that wsa setting the wrong sizes.
