+++ cratermaker/components/surface/__init__.py
@@ -875,87 +875,6 @@ class Surface(ComponentBase):
             self._save_data(uxda, interval_number, combine_data_files)
         return
 
-    @staticmethod
-    def _distribute_points(
-        distance: FloatLike,
-        radius: FloatLike = 1.0,
-        lon_range: PairOfFloats = (-180, 180),
-        lat_range: PairOfFloats = (-90, 90),
-    ) -> NDArray:
-        """
-        Distributes points on a sphere using Deserno's algorithm (Deserno 2004).
-
-        Parameters
-        ----------
-        distance : float
-            Approximate distance between points, used to determine the number of points, where n = 1/distance**2 when distributed over the whole sphere.
-        radius : float, optional
-            Radius of the sphere. Default is 1.0
-        lon_range : tuple, optional
-            Range of longitudes in degrees. Default is (-180,180).
-        lat_range : tuple, optional
-            Range of latitudes in degrees. Default is (-90,90).
-
-        Returns
-        -------
-        (3,n) ndarray of np.float64
-            Array of cartesian points on the sphere.
-
-        References
-        ----------
-        - Deserno, Markus., 2004. How to generate equidistributed points on the surface of a sphere. https://www.cmu.edu/biolphys/deserno/pdf/sphere_equi.pdf
-
-        """
-
-        def _sph2cart(theta, phi, r):
-            """
-            Converts spherical coordinates to Cartesian coordinates.
-
-            Parameters
-            ----------
-            theta : float
-                Inclination angle in radians.
-            phi : float
-                Azimuthal angle in radians.
-            r : float
-                Radius.
-            """
-            x = r * np.sin(theta) * np.cos(phi)
-            y = r * np.sin(theta) * np.sin(phi)
-            z = r * np.cos(theta)
-            return x, y, z
-
-        phi_range = np.deg2rad(lon_range) + np.pi
-        theta_range = np.deg2rad(lat_range) + np.pi / 2
-        points = []
-
-        n = int(1 / distance**2)
-        if n < 1:
-            return
-
-        a = 4 * np.pi / n
-        d = np.sqrt(a)
-        mtheta = int(np.round(np.pi / d))
-        dtheta = np.pi / mtheta
-        dphi = a / dtheta
-
-        thetavals = np.pi * (np.arange(mtheta) + 0.5) / mtheta
-        thetavals = thetavals[(thetavals >= theta_range[0]) & (thetavals < theta_range[1])]
-
-        for theta in thetavals:
-            mphi = int(np.round(2 * np.pi * np.sin(theta) / dphi))
-            phivals = 2 * np.pi * np.arange(mphi) / mphi
-            phivals = phivals[(phivals >= phi_range[0]) & (phivals < phi_range[1])]
-            for phi in phivals:
-                points.append(_sph2cart(theta, phi, radius))
-        if len(points) == 0:
-            return
-
-        points = np.array(points, dtype=np.float64)
-        points = points.T
-
-        return points
-
     @abstractmethod
     def _generate_face_distribution(self, **kwargs: Any) -> tuple[NDArray, NDArray, NDArray]: ...
